move Meet the Champions component into separate file
