+++ src/pages/community-champions.js
@@ -1,7 +1,5 @@
 import React from 'react';
 import styled from '@emotion/styled';
-import dlv from 'dlv';
-import { useStaticQuery, graphql } from 'gatsby';
 import Button from '~components/dev-hub/button';
 import CommunityChampionApplication from '~components/dev-hub/community-champion-application';
 import GreenBulletedList from '~components/dev-hub/green-bulleted-list';
@@ -9,20 +7,8 @@ import HeroBanner from '~components/dev-hub/hero-banner';
 import Layout from '~components/dev-hub/layout';
 import Link from '~components/dev-hub/link';
 import MediaBlock from '~components/dev-hub/media-block';
-import ProfileImage from '~components/dev-hub/profile-image';
 import SEO from '~components/dev-hub/SEO';
-import {
-    H1,
-    H2,
-    H3,
-    H4,
-    H5,
-    H6,
-    P,
-    P2,
-    P3,
-    P4,
-} from '~components/dev-hub/text';
+import { H1, H2, H3, H4, H5, H6, P, P2 } from '~components/dev-hub/text';
 import {
     fontSize,
     lineHeight,
@@ -49,10 +35,9 @@ import GroupImage from '~images/community-champions/group.svg';
 import RocketShipImage from '~images/community-champions/rocket-ship.svg';
 import StackSquaresImage from '~images/community-champions/stack-squares.svg';
 import PaperAndPencilImage from '~images/community-champions/paper-and-pencil.svg';
-import LocationPinIcon from '~images/community-champions/location-pin-grey.svg';
-import ChampionPlaceholderImage from '~images/community-champions/champion-placeholder.svg';
 import useMedia from '~hooks/use-media';
 import { removePathPrefixFromUrl } from '~utils/remove-path-prefix-from-url';
+import MeetTheChampions from '~components/pages/community-champions/meet-the-champions';
 
 const APPLY_BUTTON_BOTTOM_MARGIN = '14px';
 const APPLY_BUTTON_MOBILE_BOTTOM_MARGIN = '12px';
@@ -75,13 +60,6 @@ const OTHER_DETAILS_AND_REQUIREMENTS_IMAGE_SIZE = '30%';
 const OTHER_DETAILS_AND_REQUIREMENTS_IMAGE_WIDTH = '70%';
 const OTHER_DETAILS_AND_REQUIREMENTS_DESCRIPTION_MAX_WIDTH = '690px';
 const OTHER_DETAILS_AND_REQUIREMENTS_LIST_MAX_WIDTH = '480px';
-const CHAMPION_ITEM_CONTAINER_WIDTH = '200px';
-const CHAMPION_ITEM_CONTAINER_MOBILE_WIDTH = '170px';
-const CHAMPION_PROFILE_PICTURE_SIZE = 112;
-const CHAMPION_PROFILE_PICTURE_GRADIENT_OFFSET = 8;
-const LOCATION_CONTAINER_TOP_MARGIN = '4px';
-const LOCATION_CONTAINER_MOBILE_TOP_MARGIN = '2px';
-const PIN_LOCATION_IMAGE_RIGHT_MARGIN = '4px';
 const FOR_THE_FUTURE_APPLY_BUTTON_TOP_MARGIN = '48px';
 const FOR_THE_FUTURE_DESCRIPTION_MAX_WIDTH = '760px';
 const OG_IMAGE = '/public/images/champions-badge.jpg';
@@ -510,148 +488,6 @@ const OtherDetailsAndRequirements = () => (
     </OtherDetailsAndRequirementsContainer>
 );
 
-const MeetTheChampionsContainer = styled('div')`
-    background-color: ${({ theme }) => theme.colorMap.greyDarkThree};
-    margin: ${size.xlarge} -${size.xxlarge};
-    padding: ${size.xlarge} ${size.xxlarge};
-    @media ${screenSize.upToLarge} {
-        margin: ${size.large} -${size.default};
-        padding: ${size.large} ${size.default};
-    }
-`;
-
-const ChampionsContainer = styled('div')`
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;
-    row-gap: ${size.default};
-    @media ${screenSize.upToLarge} {
-        row-gap: ${size.xsmall};
-    } ;
-`;
-
-const ChampionLink = styled(Link)`
-    align-items: center;
-    border-radius: ${size.xsmall};
-    display: flex;
-    flex-direction: column;
-    padding: ${size.default};
-    text-align: center;
-    text-decoration: none;
-    width: ${CHAMPION_ITEM_CONTAINER_WIDTH};
-    @media ${screenSize.upToMedium} {
-        padding: ${size.default} 0;
-        width: ${CHAMPION_ITEM_CONTAINER_MOBILE_WIDTH};
-    }
-    &:hover,
-    &:focus {
-        background-color: ${({ theme }) => theme.colorMap.greyDarkTwo};
-        color: inherit;
-        cursor: pointer;
-    }
-`;
-
-const LocationPinImage = styled('img')`
-    margin-right: ${PIN_LOCATION_IMAGE_RIGHT_MARGIN};
-    vertical-align: text-top;
-`;
-
-const LocationText = styled(P4)`
-    color: ${({ theme }) => theme.colorMap.greyLightTwo};
-    margin-top: ${LOCATION_CONTAINER_TOP_MARGIN};
-    @media ${screenSize.upToMedium} {
-        margin-top: ${LOCATION_CONTAINER_MOBILE_TOP_MARGIN};
-    }
-`;
-
-const ChampionLocation = ({ location }) => (
-    <LocationText collapse>
-        <LocationPinImage src={LocationPinIcon} alt="Location pin" />
-        {location}
-    </LocationText>
-);
-
-const ChampionTitleText = styled(P3)`
-    color: ${({ theme }) => theme.colorMap.greyLightTwo};
-`;
-
-const ChampionProfilePicture = styled(ProfileImage)`
-    margin-bottom: ${size.default};
-    div {
-        ${props =>
-            props.image
-                ? `background-color: ${props.theme.colorMap.devWhite};`
-                : ''}
-        background-size: cover;
-    }
-`;
-
-const ChampionItem = ({ imageUrl, location, name, title, to }) => (
-    <ChampionLink to={to}>
-        <ChampionProfilePicture
-            defaultImage={ChampionPlaceholderImage}
-            gradientOffset={CHAMPION_PROFILE_PICTURE_GRADIENT_OFFSET}
-            hideOnMobile={false}
-            image={imageUrl}
-            key={name}
-            height={CHAMPION_PROFILE_PICTURE_SIZE}
-            width={CHAMPION_PROFILE_PICTURE_SIZE}
-        />
-        <H6 collapse>{name}</H6>
-        <ChampionTitleText collapse>{title}</ChampionTitleText>
-        <ChampionLocation location={location} />
-    </ChampionLink>
-);
-
-const communityChampions = graphql`
-    query CommunityChampions {
-        allStrapiCommunityChampions(
-            sort: { fields: [firstName, middleName, lastName] }
-        ) {
-            nodes {
-                id
-                firstName
-                middleName
-                lastName
-                location
-                title
-                image {
-                    url
-                }
-            }
-        }
-    }
-`;
-
-const ChampionList = () => {
-    const data = useStaticQuery(communityChampions);
-    const champions = dlv(data, ['allStrapiCommunityChampions', 'nodes'], []);
-    return (
-        <ChampionsContainer>
-            {champions.map(
-                ({
-                    firstName,
-                    id,
-                    image,
-                    lastName,
-                    location,
-                    middleName,
-                    title,
-                }) => (
-                    <ChampionItem
-                        imageUrl={image ? image.url : null}
-                        key={id}
-                        location={location}
-                        name={[firstName, middleName, lastName].join(' ')}
-                        title={title}
-                        to={`/community-champions/${firstName.toLowerCase()}-${lastName.toLowerCase()}`}
-                    />
-                )
-            )}
-        </ChampionsContainer>
-    );
-};
-
 const ForTheFutureTitle = styled(H1)`
     margin-bottom: ${size.large};
     text-align: center;
@@ -991,10 +827,7 @@ const CommunityChampions = () => {
                 </HowToQualifyContainer>
                 <Line />
                 <OtherDetailsAndRequirements />
-                <MeetTheChampionsContainer>
-                    <Title>Meet the Champions</Title>
-                    <ChampionList />
-                </MeetTheChampionsContainer>
+                <MeetTheChampions />
                 <ForTheFutureContainer>
                     <ForTheFutureTitle>For the Future</ForTheFutureTitle>
                     <ForTheFutureDescription collapse>
